% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cartel_duration.R
\name{get_cartel_duration}
\alias{get_cartel_duration}
\title{Cartel Duration}
\usage{
get_cartel_duration(cartel_dates, detect_dates, leniency_dates)
}
\arguments{
\item{cartel_dates}{binary collusive state}

\item{detect_dates}{binary collusive state that eventually got detected}

\item{leniency_dates}{binary collusive state that eventually got found due to leniency application}
}
\value{
cross-sectional dataset with duration
}
\description{
Function to combine paneldata cartels, detected cartels, leniency cases returned by function sim_col into a cross-sectional dataset with duration
}
\examples{
sim_list <- sim_col();
df <- get_cartel_duration(sim_list$cartels, sim_list$detection, sim_list$leniency);
}
