% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_collusion.R
\name{sim_col}
\alias{sim_col}
\title{Simulate Collusion}
\usage{
sim_col(
  model = "1",
  periods = 100,
  n_ind = 100,
  delta = 0.85,
  pi_mean = 0,
  pi_sd = 1.5,
  pi_min = 1,
  pi_max = 100,
  k_min = 0,
  k_max = 0.1,
  eta_mean = 1,
  eta_sd = 1.5,
  eta_min = 1.1,
  eta_max = 10,
  S = 0.8,
  Q = 0.2,
  R = 0.3,
  gamma = 2,
  theta = 1,
  prob_leniency = 1,
  lambda = 0.5,
  tau = 0.8,
  xi = 1,
  v = 500,
  rho = 1.5
)
}
\arguments{
\item{model}{Modelname (1, 2, 2B, 3)}

\item{periods}{Number of time periods}

\item{n_ind}{Number of industries}
}
\value{
List with panel data cartels, detected cartels, leniency cases, profit over time, time varying parameters, ICC over time
}
\description{
Simulation of industries with collusive and non-collusive periods over time, detected and undetected, based on Harrington/Chang (2015)
}
\examples{
sim_list <- sim_col();
sim_list <- sim_col(model = "2", periods = 1000, n_ind = 1000, theta = 0.035);
sim_list <- sim_col(model = "2B", periods = 100, n_ind = 100, theta = 0.035);
sim_list <- sim_col(model = "3", periods = 10, n_ind = 10);
}
